Update ContactBook.py
Made a file where numerous dictionaries are stored. In each dictionary in the file, a username and password is held in the form: -> key:value. 
Now, a user can login using their credentials even if they registered before and the program is run again. This means that the data is being saved into the file so that user details can be accessed
Did some error checking so that if the user doesn't enter the correct credentials, they get an error message. Also removed some parameters that I passed into the log_on_system() function.