Update ContactBook.py
I changed the program so that when i add, remove or view a contact, it will open an individual file for each user so that one user cannot see the contacts of the other user. 
Also, if the user wants to add a contact and their account database is empty (they have just registered or haven't used the system yet), a table will be created so that the user can then put details of their contacts such as name, address, phone number and email.