Update ContactBook.py
Added an option to exit the program. This means that if the user wants to quit using the program/app, they can just choose that option. Also, added some error checking to the update contact function so that the user cannot update the contacts if there are no contacts available