Create DicePics.md
These are the pictures I used for the dice simulation. You need them in the dice simulation project, if you wish to run the program. You have to put them in the directory in which the project is located.  The final picture shows the logo picture that I used  It is in the .png format. You need to change it to .ico. GitHub does not allow me to upload the .ico format. There are many website online that convert this easily